Prep HTTP layer marshaling code
