+++ http/marshal.go
@@ -1,0 +1,22 @@
+package http
+
+// Result represents the outcome of an operation that changes rows.
+type Result struct {
+	LastInsertID int64   `json:"last_insert_id,omitempty"`
+	RowsAffected int64   `json:"rows_affected,omitempty"`
+	Error        string  `json:"error,omitempty"`
+	Time         float64 `json:"time,omitempty"`
+}
+
+// Rows represents the outcome of an operation that returns query data.
+type Rows struct {
+	Columns []string        `json:"columns,omitempty"`
+	Types   []string        `json:"types,omitempty"`
+	Values  [][]interface{} `json:"values,omitempty"`
+	Error   string          `json:"error,omitempty"`
+	Time    float64         `json:"time,omitempty"`
+}
+
+func Convert(r []*command.ExecuteResult) *Result {
+	return nil
+}
